+++ odtp/dashboard/page_executions/main.py
@@ -9,7 +9,6 @@ import odtp.mongodb.db as db
 import odtp.dashboard.page_executions.table as table
 import odtp.dashboard.page_executions.select as select
 import odtp.dashboard.page_executions.add as add
-import odtp.dashboard.page_executions.add2 as add2
 import odtp.dashboard.page_executions.workarea as workarea
 
 log = logging.getLogger("__name__")
@@ -49,35 +48,23 @@ def parse_key_value_pairs(text: str) -> dict:
 def ui_tabs(current_digital_twin, workdir):
     try:
         with ui.tabs() as tabs:
-            select = ui.tab("Select an execution")
+            table = ui.tab("Manage execution")
+            select = ui.tab("Execution Details")
             add = ui.tab("Add an execution")
-            table = ui.tab("Execution table")
         with ui.tab_panels(tabs, value=select).classes("w-full") as panels:
             with ui.tab_panel(select):
                 ui_execution_select(current_digital_twin)
                 ui_execution_details()
             with ui.tab_panel(add):
-                #ui_add_execution(current_digital_twin, workdir)
-                ui_add_execution2(current_digital_twin)
+                ui_add_execution(current_digital_twin)
             with ui.tab_panel(table):
                 ui_executions_table(current_digital_twin)
     except Exception as e:
         log.exception(f"Execution Tabs could not be loaded. An Exception occurred: {e}")
 
 
-@ui.refreshable
-def ui_add_execution(current_digital_twin, workdir):
-    try:
-        current_execution_to_add = storage.get_active_object_from_storage(
-            storage.NEW_EXECUTION
-        )
-        add.ui_execution_add_form(current_digital_twin, workdir, current_execution_to_add)
-    except Exception as e:
-        log.exception(f"Execution add form could not be loaded. An Exception occurred: {e}")
-
-
 def ui_add_execution2(current_digital_twin):
-    add2.ExecutionForm(current_digital_twin["digital_twin_id"])
+    add.ExecutionForm(current_digital_twin["digital_twin_id"])
 
 
 @ui.refreshable
